refactored dependency injection

+++ src/main/java/at/fhtw/swen3/persistence/repositories/WarehouseRepository.java
@@ -4,6 +4,8 @@ import at.fhtw.swen3.persistence.entities.WarehouseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface WarehouseRepository extends JpaRepository<WarehouseEntity, Integer> {
+public interface WarehouseRepository extends JpaRepository<WarehouseEntity, Long> {
 }
